Huge Frontend Overhaul
