+++ Quizard_client/src/componenets/Layout.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import { ThemeToggle } from './ThemeToggle';
+
+interface LayoutProps {
+  children: React.ReactNode;
+}
+
+export const Layout: React.FC<LayoutProps> = ({ children }) => (
+  <div className="app-container">
+    <ThemeToggle />
+    {children}
+  </div>
+);
