+++ Quizard_client/src/main.tsx
@@ -1,10 +1,14 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
+import { auth0Config } from './auth0-config';
+import { Auth0Provider } from '@auth0/auth0-react';
 import './index.css'
 import App from './App.tsx'
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
-    <App />
+    <Auth0Provider {...auth0Config}>
+      <App />
+    </Auth0Provider>
   </StrictMode>,
 )
